Fixes to sepparate server and application
