+++ internal/delivery/http/server.go
@@ -1,13 +1,28 @@
 package http
 
 import (
+	"encoding/json"
+	"fmt"
+	"io"
 	"net/http"
+	"strconv"
+	"time"
 
 	_ "net/http/pprof"
 
 	"github.com/prometheus/client_golang/prometheus/promhttp"
+	log "github.com/sirupsen/logrus"
 	"github.com/tentens-tech/shared-lock/internal/application"
+	"github.com/tentens-tech/shared-lock/internal/application/command/leasemanagement"
 	"github.com/tentens-tech/shared-lock/internal/config"
+	"github.com/tentens-tech/shared-lock/internal/infrastructure/cache"
+	"github.com/tentens-tech/shared-lock/internal/infrastructure/metrics"
+	"github.com/tentens-tech/shared-lock/internal/infrastructure/storage"
+)
+
+const (
+	defaultLeaseTTLHeader = "x-lease-ttl"
+	defaultLeaseDuration  = 10 * time.Second
 )
 
 type Server struct {
@@ -27,7 +42,10 @@ func (s *Server) Start(cfg *config.ServerCfg) error {
 	mux.HandleFunc("/keepalive", s.handleKeepalive)
 	mux.HandleFunc("/health", s.handleHealth)
 	mux.Handle("/metrics", promhttp.Handler())
-	mux.HandleFunc("/debug/pprof/", http.HandlerFunc(http.DefaultServeMux.ServeHTTP))
+
+	if cfg.PPROFEnabled {
+		mux.HandleFunc("/debug/pprof/", http.HandlerFunc(http.DefaultServeMux.ServeHTTP))
+	}
 
 	s.Server = &http.Server{
 		Addr:         ":" + cfg.Port,
@@ -41,13 +59,107 @@ func (s *Server) Start(cfg *config.ServerCfg) error {
 }
 
 func (s *Server) handleLease(w http.ResponseWriter, r *http.Request) {
-	application.GetLeaseHandler(s.app.Ctx, s.app.Config, s.app.StorageConnection, s.app.LeaseCache)(w, r)
+	start := time.Now()
+	defer func() {
+		metrics.LeaseOperationDuration.WithLabelValues(metrics.LeaseOperationGet).Observe(time.Since(start).Seconds())
+	}()
+
+	var err error
+	var lease leasemanagement.Lease
+	var leaseID int64
+	var leaseStatus string
+
+	body, err := io.ReadAll(r.Body)
+	if err != nil {
+		log.Errorf("Failed to read request body, %v", err)
+		http.Error(w, "Failed to read request body", http.StatusBadRequest)
+		return
+	}
+
+	log.Debugf("Request body: %v", string(body))
+	err = json.Unmarshal(body, &lease)
+	if err != nil {
+		log.Errorf("Failed to unmarshal request body, %v", err)
+		http.Error(w, "Failed to unmarshal request body", http.StatusBadRequest)
+		return
+	}
+
+	if s.app.LeaseCache != nil {
+		if cachedValue, exists := s.app.LeaseCache.Get(lease.Key); exists {
+			if cachedLease, ok := cachedValue.(cache.LeaseCacheRecord); ok {
+				leaseStatus = cachedLease.Status
+				leaseID = cachedLease.ID
+				log.Debugf("Cache hit for lease key: %v", lease.Key)
+				metrics.CacheOperations.WithLabelValues(metrics.LeaseOperationGet, "success").Inc()
+			}
+		}
+	}
+
+	if leaseStatus == "" {
+		leaseTTL, err := time.ParseDuration(r.Header.Get(defaultLeaseTTLHeader))
+		if err != nil {
+			log.Warnf("Can't parse value of %v header. Using defaultLeaseDuration for %v", defaultLeaseTTLHeader, lease.Key)
+			leaseTTL = defaultLeaseDuration
+		}
+
+		leaseStatus, leaseID, err = s.app.CreateLease(leaseTTL, lease)
+		if err != nil {
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+
+		if s.app.LeaseCache != nil {
+			s.app.LeaseCache.Set(lease.Key, cache.LeaseCacheRecord{
+				Status: leaseStatus,
+				ID:     leaseID,
+			}, leaseTTL)
+		}
+	}
+
+	switch leaseStatus {
+	case storage.StatusAccepted:
+		w.WriteHeader(http.StatusAccepted)
+	case storage.StatusCreated:
+		w.WriteHeader(http.StatusCreated)
+	default:
+		w.WriteHeader(http.StatusInternalServerError)
+	}
+
+	_, err = w.Write([]byte(fmt.Sprintf("%v", leaseID)))
+	if err != nil {
+		log.Errorf("Failed to write response for /lease endpoint, %v", err)
+		return
+	}
 }
 
 func (s *Server) handleKeepalive(w http.ResponseWriter, r *http.Request) {
-	application.KeepaliveHandler(s.app.Ctx, s.app.Config, s.app.StorageConnection)(w, r)
+	var err error
+	var leaseID int64
+
+	body, err := io.ReadAll(r.Body)
+	if err != nil {
+		log.Errorf("Failed to read request body, %v", err)
+		http.Error(w, "Failed to read request body", http.StatusBadRequest)
+		return
+	}
+
+	log.Debugf("Request body: %v", string(body))
+	leaseID, err = strconv.ParseInt(string(body), 10, 64)
+	if err != nil {
+		log.Errorf("Failed to parse lease id from string, leaseIDString: %v, %v", string(body), err)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	log.Debugf("Trying to revive lease: %v", leaseID)
+	err = s.app.ReviveLease(leaseID)
+	if err != nil {
+		log.Warnf("Failed to prolong lease: %v", err)
+		http.Error(w, "Failed to prolong lease", http.StatusNoContent)
+		return
+	}
 }
 
 func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
-	application.HealthHandler()(w, r)
+	w.WriteHeader(http.StatusOK)
 }
